Backend/app/main.py
from dotenv import load_dotenv
load_dotenv()
# app/main.py
from fastapi import FastAPI
from .database import Base, engine
from .routes import leads, ai, outreach, replies, proposals, payments, auth, email_logs, stripe_webhook
from .routes import users
from fastapi.middleware.cors import CORSMiddleware
from app.routes import webhooks


Base.metadata.create_all(bind=engine)

app = FastAPI(title="CRM AutoPilot")

app.include_router(leads.router)
app.include_router(ai.router)
app.include_router(outreach.router)
app.include_router(replies.router)
app.include_router(proposals.router)
app.include_router(payments.router)
app.include_router(auth.router)
app.include_router(users.router)
app.include_router(email_logs.router)
app.include_router(stripe_webhook.router)

app.include_router(webhooks.router)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/")
def root():
    return {"message": "CRM AutoPilot Backend Running"}



# from fastapi import FastAPI

# app = FastAPI()

# @app.get("/")
# def root():
#     return {"message": "Server OK"}



# from fastapi import FastAPI
# from .database import Base, engine

# Base.metadata.create_all(bind=engine)

# app = FastAPI()

# @app.get("/")
# def root():
#     return {"message": "DB OK"}





# from fastapi import FastAPI
# from .database import Base, engine
# from .routes import leads

# Base.metadata.create_all(bind=engine)

# app = FastAPI()
# app.include_router(leads.router)

# @app.get("/")
# def root():
#     return {"message": "Leads OK"}